make `temperature` a float

+++ infra/migrations/1726598024115_create-data-table.js
@@ -31,7 +31,7 @@ exports.up = (pgm) => {
     },
 
     temperature: {
-      type: "integer",
+      type: "float",
     },
 
     light_intensity: {
